Tests/Docs after stricter explicit conversion.

+++ test/libsolidity/SolidityEndToEndTest.cpp
@@ -675,7 +675,7 @@ BOOST_AUTO_TEST_CASE(sign_extension)
 				unchecked {
 					int64 x = -int32(0xff);
 					if (x >= 0xff) return 0;
-					return 0 - uint256(x);
+					return 0 - uint64(x);
 				}
 			}
 		}
@@ -3194,7 +3194,7 @@ BOOST_AUTO_TEST_CASE(byte_array_pop_long_storage_empty)
 				for (uint8 i = 0; i <= 40; i++)
 					data.push(byte(i+1));
 				for (int8 j = 40; j >= 0; j--) {
-					require(data[uint8(j)] == byte(j+1));
+					require(data[uint8(j)] == byte(uint8(j+1)));
 					require(data.length == uint8(j+1));
 					data.pop();
 				}
@@ -6575,7 +6575,7 @@ BOOST_AUTO_TEST_CASE(dirty_scratch_space_prior_to_constant_optimiser)
 				return 0x0000000000001234123412431234123412412342112341234124312341234124;
 			}
 			function g(address a) internal pure returns (uint) {
-				unchecked { return uint(a) * 0x0000000000001234123412431234123412412342112341234124312341234124; }
+				unchecked { return uint160(a) * 0x0000000000001234123412431234123412412342112341234124312341234124; }
 			}
 			function h(uint a) internal pure returns (uint) {
 				unchecked { return a * 0x0000000000001234123412431234123412412342112341234124312341234124; }
